refactor: my-book-list を my-book に移動
