--- feature/my-book-list/src/main/java/app/kaito_dogi/mybrary/feature/mybooklist/MyBookListNavigation.kt
@@ -1,18 +1,0 @@
-package app.kaito_dogi.mybrary.feature.mybooklist
-
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import app.kaito_dogi.mybrary.core.domain.model.MyBook
-import app.kaito_dogi.mybrary.core.ui.navigation.MybraryRoute
-
-fun NavGraphBuilder.myBookListScreen(
-  onAdditionClick: () -> Unit,
-  onMyBookClick: (MyBook) -> Unit,
-) {
-  composable<MybraryRoute.MyBookList> {
-    MyBookListScreenContainer(
-      onAdditionClick = onAdditionClick,
-      onMyBookClick = onMyBookClick,
-    )
-  }
-}
