+++ feature/my-book/src/main/java/app/kaito_dogi/mybrary/feature/mybook/destination/mybooklist/component/MyBookListHeader.kt
@@ -1,14 +1,15 @@
-package app.kaito_dogi.mybrary.feature.mybooklist.component
+package app.kaito_dogi.mybrary.feature.mybook.destination.mybooklist.component
 
+import androidx.annotation.StringRes
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import app.kaito_dogi.mybrary.core.designsystem.component.SkeletonBox
 import app.kaito_dogi.mybrary.core.designsystem.component.Text
 import app.kaito_dogi.mybrary.core.designsystem.theme.MybraryTheme
 import app.kaito_dogi.mybrary.core.ui.R
@@ -16,27 +17,30 @@ import app.kaito_dogi.mybrary.core.ui.R
 private val Height = 40.dp
 
 @Composable
-internal fun MyBookListHeaderSkeleton(
+internal fun MyBookListHeader(
+  @StringRes titleResId: Int,
   modifier: Modifier = Modifier,
 ) {
-  SkeletonBox(
-    shape = MybraryTheme.shapes.extraSmall,
+  Column(
     modifier = modifier
       .fillMaxWidth()
       .height(Height),
+    verticalArrangement = Arrangement.Center,
   ) {
     Text(
-      text = stringResource(id = R.string.my_book_list_text_a),
-      color = Color.Transparent,
+      text = stringResource(id = titleResId),
+      color = MybraryTheme.colorScheme.onBackground,
       style = MybraryTheme.typography.titleLarge,
     )
   }
 }
 
-@Preview
+@Preview(showBackground = true)
 @Composable
-private fun MyBookListHeaderSkeletonPreview() {
+private fun MyBookListHeaderPreview() {
   MybraryTheme {
-    MyBookListHeaderSkeleton()
+    MyBookListHeader(
+      titleResId = R.string.my_book_list_text_all_my_books,
+    )
   }
 }
