+++ feature/my-book/src/main/java/app/kaito_dogi/mybrary/feature/mybook/destination/mybooklist/component/MyBookListHeaderSkeleton.kt
@@ -1,15 +1,14 @@
-package app.kaito_dogi.mybrary.feature.mybooklist.component
+package app.kaito_dogi.mybrary.feature.mybook.destination.mybooklist.component
 
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import app.kaito_dogi.mybrary.core.designsystem.component.SkeletonBox
 import app.kaito_dogi.mybrary.core.designsystem.component.Text
 import app.kaito_dogi.mybrary.core.designsystem.theme.MybraryTheme
 import app.kaito_dogi.mybrary.core.ui.R
@@ -17,30 +16,27 @@ import app.kaito_dogi.mybrary.core.ui.R
 private val Height = 40.dp
 
 @Composable
-internal fun MyBookListHeader(
-  @StringRes titleResId: Int,
+internal fun MyBookListHeaderSkeleton(
   modifier: Modifier = Modifier,
 ) {
-  Column(
+  SkeletonBox(
+    shape = MybraryTheme.shapes.extraSmall,
     modifier = modifier
       .fillMaxWidth()
       .height(Height),
-    verticalArrangement = Arrangement.Center,
   ) {
     Text(
-      text = stringResource(id = titleResId),
-      color = MybraryTheme.colorScheme.onBackground,
+      text = stringResource(id = R.string.my_book_list_text_a),
+      color = Color.Transparent,
       style = MybraryTheme.typography.titleLarge,
     )
   }
 }
 
-@Preview(showBackground = true)
+@Preview
 @Composable
-private fun MyBookListHeaderPreview() {
+private fun MyBookListHeaderSkeletonPreview() {
   MybraryTheme {
-    MyBookListHeader(
-      titleResId = R.string.my_book_list_text_all_my_books,
-    )
+    MyBookListHeaderSkeleton()
   }
 }
