+++ settings.gradle.kts
@@ -27,7 +27,6 @@ include(":core:supabase")
 include(":core:ui")
 include(":feature:my-book")
 include(":feature:my-book-detail")
-include(":feature:my-book-list")
 include(":feature:search-book")
 include(":feature:send-otp")
 include(":feature:setting")
